made a few updates based on comments

+++ cordova-lib/src/plugman/registry/registry.js
@@ -89,8 +89,14 @@ module.exports = {
         .then(function(settings) {
             if(fs.existsSync(path.join(dir,'package.json'))) {
                 events.emit('verbose', 'temporarily moving existing package.json so we can create one to publish to the cordova plugins registry');
-                //rename package.json to pacakge.json1 temporarily 
-                fs.renameSync(path.join(dir,'package.json'),path.join(dir,'package.json1')); 
+                if(fs.existsSync(path.join(dir,'package.json1'))) {
+                    //package.json1 already exists, maybe due to an past failed attempt to publish
+                    //we will assume that the rename has already happened.
+                    events.emit('verbose', 'package.json1 already exists. Will use');
+                } else {
+                    //rename package.json to pacakge.json1 temporarily 
+                    fs.renameSync(path.join(dir,'package.json'),path.join(dir,'package.json1')); 
+                }
             }
             return manifest.generatePackageJsonFromPluginXml(dir)
             .then(function() {
